+++ Fluid.Tests/TemplateTests.cs
@@ -2,13 +2,11 @@
 using System.Collections.Generic;
 using System.Globalization;
 using System.IO;
-using System.Text;
 using System.Text.Encodings.Web;
 using System.Threading.Tasks;
 using Fluid.Tests.Domain;
+using Fluid.Tests.Mocks;
 using Fluid.Values;
-using Microsoft.Extensions.FileProviders;
-using Microsoft.Extensions.Primitives;
 using Xunit;
 
 namespace Fluid.Tests
@@ -381,9 +379,10 @@ namespace Fluid.Tests
             var expectedUS = "1234.567";
 
             FluidTemplate.TryParse(source, out var template, out var messages);
-            var context = new TemplateContext();
-
-            context.CultureInfo = new CultureInfo("en-US");
+            var context = new TemplateContext
+            {
+                CultureInfo = new CultureInfo("en-US")
+            };
             var resultUS = await template.RenderAsync(context);
 
             context.CultureInfo = new CultureInfo("fr-FR");
@@ -460,7 +459,7 @@ shape: 'circle'";
             FluidTemplate.TryParse(source, out var template, out var messages);
             var context = new TemplateContext
             {
-                FileProvider = new TestFileProvider("Partials")
+                FileProvider = new MockFileProvider("Partials")
             };
             var result = await template.RenderAsync(context);
 
@@ -479,72 +478,11 @@ shape: ''value";
             FluidTemplate.TryParse(source, out var template, out var messages);
             var context = new TemplateContext
             {
-                FileProvider = new TestFileProvider("Partials")
+                FileProvider = new MockFileProvider("Partials")
             };
             var result = await template.RenderAsync(context);
 
             Assert.Equal(expected, result);
-        }
-
-        public class TestFileProvider : IFileProvider
-        {
-            private string _partialsFolderPath;
-
-            public TestFileProvider(string path)
-            {
-                if (path != "Partials")
-                {
-                    throw new DirectoryNotFoundException();
-                }
-
-                _partialsFolderPath = path;
-            }
-
-            public IDirectoryContents GetDirectoryContents(string subpath)
-            {
-                throw new NotImplementedException();
-            }
-
-            public IFileInfo GetFileInfo(string subpath)
-            {
-                var path = Path.Combine(_partialsFolderPath, subpath);
-                return new TestFileInfo(path);
-            }
-
-            public IChangeToken Watch(string filter)
-            {
-                throw new NotImplementedException();
-            }
-        }
-
-        public class TestFileInfo : IFileInfo
-        {
-            public TestFileInfo(string name)
-            {
-                Name = name;
-            }
-
-            public bool Exists => true;
-
-            public bool IsDirectory => false;
-
-            public DateTimeOffset LastModified => DateTimeOffset.MinValue;
-
-            public long Length => -1;
-
-            public string Name { get; }
-
-            public string PhysicalPath => null;
-
-            public Stream CreateReadStream()
-            {
-                var content = @"{{ 'Partial Content' }}
-color: '{{ color }}'
-shape: '{{ shape }}'";
-                var data = Encoding.UTF8.GetBytes(content);
-
-                return new MemoryStream(data);
-            }
         }
     }
 }
